Restructure redux and store state between routes
